Final fix comments and fields

+++ src/controllers/fields.controller.js
@@ -2,6 +2,8 @@ import fieldModel from "../models/fields.model.js";
 import validation from "../helpers/fields.validation.js";
 import { uploadImage, deleteImage } from "../libs/cloudinary.js";
 import fs from "fs-extra";
+import sportCenterModel from "../models/sportCenter.model.js";
+import { createError } from "../utils/error.js";
 
 //GET
 const getAllFields = async (req, res) => {
@@ -15,13 +17,13 @@ const getAllFields = async (req, res) => {
 };
 
 //GET by ID
-const getFieldByID = async (req, res) => {
+const getFieldByID = async (req, res, next) => {
   try {
     const field = await fieldModel.findById(req.params.id);
     if (field) {
       res.status(200).json(field);
     } else {
-      res.status(404).json("Field Not Found");
+      return next(createError(404, "Cancha no encontrada"));
     }
   } catch (error) {
     console.log(error);
@@ -45,7 +47,7 @@ const getPage = async (req, res) => {
 };
 
 //POST
-const createField = async (req, res) => {
+const createField = async (req, res, next) => {
   //Only for the owner of the sportCenter, validate that
   try {
     const bodyfield = {
@@ -57,36 +59,46 @@ const createField = async (req, res) => {
       isActive: req.body.isActive,
       idSportCenter: req.body.idSportCenter,
     };
-    console.log(
-      validation.hourValidation(bodyfield.openHour, bodyfield.closeHour)
-    );
 
-    if (!validation.createFieldDataValidation(bodyfield)) {
-      res.status(400).json("Some data is missing");
-    } else if (
-      validation.nameValidation(bodyfield.name) &&
-      validation.hourValidation(bodyfield.openHour, bodyfield.closeHour) &&
-      validation.priceValidation(bodyfield.pricePerHour) &&
-      validation.sizeValidation(bodyfield.size) &&
-      (await validation.validateSportCenter(bodyfield.idSportCenter))
-    ) {
-      const photo = {
-        url: "",
-        public_id: "",
-      };
-      if (req.files.image) {
-        console.log(req.files.image);
-        const result = await uploadImage(req.files.image.tempFilePath);
-        photo.url = result.secure_url;
-        photo.public_id = result.public_id;
+    if(await validation.validateSportCenter(req.body.idSportCenter)){
+      const sportCenter = await sportCenterModel.findById(bodyfield.idSportCenter);
+      if((req.user.id == sportCenter.ownerId) || req.user.isAdmin){
+        if (!validation.createFieldDataValidation(bodyfield)) {
+          return next(createError(400, "Falta ingresar información"));
+        } else if (
+          validation.nameValidation(bodyfield.name) &&
+          validation.hourValidation(bodyfield.openHour, bodyfield.closeHour) &&
+          validation.priceValidation(bodyfield.pricePerHour) &&
+          validation.sizeValidation(bodyfield.size) &&
+          (await validation.validateSportCenter(bodyfield.idSportCenter))
+        ) {
+          const photo = {
+            url: "",
+            public_id: "",
+          };
+          if (req.files.image) {
+            console.log(req.files.image);
+            const result = await uploadImage(req.files.image.tempFilePath);
+            photo.url = result.secure_url;
+            photo.public_id = result.public_id;
+          }
+          const newField = new fieldModel({ ...bodyfield, photo });
+          await newField.save();
+          fs.remove(req.files.image.tempFilePath);
+          res.status(201).json(newField);
+        } else {
+          return next(createError(404, "Informacion no válida"));
+        }
       }
-      const newField = new fieldModel({ ...bodyfield, photo });
-      await newField.save();
-      fs.remove(req.files.image.tempFilePath);
-      res.status(201).json(newField);
-    } else {
-      res.status(400).json("The written data is invalid");
+      else {
+        return next(createError(400, "No esta autorizado a crear canchas en este complejo"));
+      }
     }
+    else {
+      return next(createError(400, "Complejo no válido"));
+    }
+
+
   } catch (error) {
     console.log(error);
     res.status(500).json({ message: error.message });
@@ -94,43 +106,49 @@ const createField = async (req, res) => {
 };
 
 //PUT
-const updateField = async (req, res) => {
+const updateField = async (req, res, next) => {
   //For activate or deactivate a field, update that
   //Only for the owner of the sportCenter, validate that
   try {
     const field = await fieldModel.findById(req.params.id);
     if (field) {
-      if (req.body.name) field.name = req.body.name;
-      if (req.body.openHour) field.openHour = req.body.openHour;
-      if (req.body.closeHour) field.closeHour = req.body.closeHour;
-      if (req.body.pricePerHour) field.pricePerHour = req.body.pricePerHour;
-      if (req.body.size) field.size = req.body.size;
-      if (req.body.isActive) field.isActive = req.body.isActive;
-
-      if (
-        validation.nameValidation(field.name) &&
-        validation.hourValidation(field.openHour, field.closeHour) &&
-        validation.priceValidation(field.pricePerHour) &&
-        validation.sizeValidation(field.size)
-      ) {
-        if (req.files && req.files.image) {
-          const result = await uploadImage(req.files.image.tempFilePath);
-
-          if (field.photo.public_id) {
-            await deleteImage(field.photo.public_id);
+      const sportCenter = await sportCenterModel.findById(field.idSportCenter);
+      if((sportCenter.ownerId == req.user.id) || req.user.isAdmin) {
+        if (req.body.name) field.name = req.body.name;
+        if (req.body.openHour) field.openHour = req.body.openHour;
+        if (req.body.closeHour) field.closeHour = req.body.closeHour;
+        if (req.body.pricePerHour) field.pricePerHour = req.body.pricePerHour;
+        if (req.body.size) field.size = req.body.size;
+        if (req.body.isActive) field.isActive = req.body.isActive;
+  
+        if (
+          validation.nameValidation(field.name) &&
+          validation.hourValidation(field.openHour, field.closeHour) &&
+          validation.priceValidation(field.pricePerHour) &&
+          validation.sizeValidation(field.size)
+        ) {
+          if (req.files && req.files.image) {
+            const result = await uploadImage(req.files.image.tempFilePath);
+  
+            if (field.photo.public_id) {
+              await deleteImage(field.photo.public_id);
+            }
+  
+            field.photo.url = result.secure_url;
+            field.photo.public_id = result.public_id;
+            fs.remove(req.files.image.tempFilePath);
           }
-
-          field.photo.url = result.secure_url;
-          field.photo.public_id = result.public_id;
-          fs.remove(req.files.image.tempFilePath);
+          await field.save();
+          res.status(200).json(field);
+        } else {
+          return next(createError(400, "Información inválida"));
         }
-        await field.save();
-        res.status(200).json(field);
-      } else {
-        res.status(400).json("The written data is invalid");
+      }
+      else {
+        return next(createError(400, "No está autorizado a modificar canchas en este complejo"));
       }
     } else {
-      res.status(404).json("Field Not Found");
+      return next(createError(404, "Cancha no encontrada"));
     }
   } catch (error) {
     console.log(error);
@@ -139,21 +157,33 @@ const updateField = async (req, res) => {
 };
 
 //DELETE
-const deleteField = async (req, res) => {
+const deleteField = async (req, res, next) => {
   //Only for the owner of the sportCenter, validate that
   try {
-    const deletedField = await fieldModel.findOneAndDelete({
-      _id: { $eq: req.params.id },
-    }); //$eq: req.params.id means that the id from params must be equa to the field's id
-    if (deletedField) {
-      if (deletedField.photo.public_id) {
-        await deleteImage(deletedField.photo.public_id);
+    const field = await fieldModel.findById(req.params.id);
+    if(field){
+      const sportCenter = await sportCenterModel.findById(field.idSportCenter);
+      if((sportCenter.ownerId == req.user.id) || req.user.isAdmin){
+        const deletedField = await fieldModel.findOneAndDelete({
+          _id: { $eq: req.params.id },
+        }); //$eq: req.params.id means that the id from params must be equa to the field's id
+        if (deletedField) {
+          if (deletedField.photo && deletedField.photo.public_id) {
+            await deleteImage(deletedField.photo.public_id);
+          }
+          res
+            .status(200)
+            .json({ message: "The Field has been Deleted", field: deletedField });
+        } else {
+          return next(createError(404, "Cancha no encontrada"));
+        }
       }
-      res
-        .status(200)
-        .json({ message: "The Field has been Deleted", field: deletedField });
-    } else {
-      res.status(404).send("Field Not Found");
+      else{
+        return next(createError(400, "No está autorizado a eliminar canchas en este complejo"));
+      }
+    }
+    else{
+      return next(createError(404, "Cancha no encontrada"));
     }
   } catch (error) {
     console.log(error);
